Refactor: reorganize files and structure modules
