+++ lua/docscribe/init.lua
@@ -1,17 +1,23 @@
 local config = require("docscribe.config")
-local commands = require("docscribe.core.commands")
+local docscribe_generate = require("docscribe.commands.generate_docs").generate_docs_command
 
 local M = {}
 
-function M.setup(user_config)
-    config.setup(user_config)
-
-    vim.api.nvim_create_user_command("DocscribeGenerate", commands.generate_docs_for_function_under_cursor, {})
-
+local function set_metadata()
     local highlight_color = config.get_config("ui").highlight.bg
     if highlight_color then
         vim.cmd("highlight DocscribeProcessing guibg=" .. highlight_color)
     end
+end
+
+local function set_commands()
+    vim.api.nvim_create_user_command("DocscribeGenerate", docscribe_generate, {})
+end
+
+function M.setup(user_config)
+    config.setup(user_config)
+    set_metadata()
+    set_commands()
 end
 
 return M
