+++ lua/docscribe/prompt_templates/init.lua
@@ -1,0 +1,10 @@
+return {
+	default_template = require("docscribe.prompt_templates.default"),
+	c_template = require("docscribe.prompt_templates.c"),
+	cpp_template = require("docscribe.prompt_templates.cpp"),
+	python_template = require("docscribe.prompt_templates.python"),
+	java_template = require("docscribe.prompt_templates.java"),
+	lua_template = require("docscribe.prompt_templates.lua"),
+	javascript_template = require("docscribe.prompt_templates.javascript"),
+	typescript_template = require("docscribe.prompt_templates.typescript"),
+}
